pyproject.toml
[tool.poetry]
name = "lexnet"
version = "0.1.0"
description = ""
authors = ["Viraat Das <viraat.laldas@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
graphiti-core = {path = "/Users/viraatd/Documents/Personal/graphiti", develop = true}
llama-index = "^0.11.5"
llama-index-llms-groq = "^0.2.0"
llama-index-embeddings-huggingface = "^0.3.1"
llama-parse = "^0.5.2"
python-dotenv = "^1.0.1"
groq="^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
